MI-206 Added the distance travelled metrics to the db
Added the distance travelled metrics to the Ride table in the database. For each ride, there will be a distance travelled value thatwill be stored when the ride is over.
Also added method to set the distance travelled metric, to facilitate use from the application-side. It will allow the developers to quickly set the distance travelled value at the end of a ride.