MI-207 Added duration of travel metric to db
Added duration of travel metric to the database, Ride table. It will be stored as a long attribute of the duration of the ride in seconds.
Also added the respective code to update this value in the database in the ride controller and ride catalog, with their respective documentation (XML comments).